+++ test/lib/zeitwerk/test_push_dir.rb
@@ -9,11 +9,22 @@ class TesPushDir < LoaderTest
   def check_dirs
     roots = loader.__roots
 
+    non_ignored_roots = roots.reject { |dir, _| loader.send(:ignored_path?, dir) }
+
     dirs = loader.dirs
-    assert_equal roots.keys, dirs
+    assert_equal non_ignored_roots.keys, dirs
     assert dirs.frozen?
 
     dirs = loader.dirs(namespaces: true)
+    assert_equal non_ignored_roots, dirs
+    assert dirs.frozen?
+    assert !dirs.equal?(roots)
+
+    dirs = loader.dirs(ignored: true)
+    assert_equal roots.keys, dirs
+    assert dirs.frozen?
+
+    dirs = loader.dirs(namespaces: true, ignored: true)
     assert_equal roots, dirs
     assert dirs.frozen?
     assert !dirs.equal?(roots)
@@ -29,6 +40,24 @@ class TesPushDir < LoaderTest
     check_dirs
   end
 
+  test "there can be several root directories" do
+    files = [["rd1/x.rb", "X = 1"], ["rd2/y.rb", "Y = 1"], ["rd3/z.rb", "Z = 1"]]
+    with_setup(files) do
+      check_dirs
+    end
+  end
+
+  test "there can be several root directories, some of them may be ignored" do
+    files = [["rd1/x.rb", "X = 1"], ["rd2/y.rb", "Y = 1"], ["rd3/z.rb", "Z = 1"]]
+    with_files(files) do
+      loader.push_dir("rd1")
+      loader.push_dir("rd2")
+      loader.push_dir("rd3")
+      loader.ignore("rd2")
+      check_dirs
+    end
+  end
+
   test "accepts dirs as strings and associates them to the given namespace" do
     loader.push_dir(".", namespace: Namespace)
     check_dirs
@@ -37,6 +66,24 @@ class TesPushDir < LoaderTest
   test "accepts dirs as pathnames and associates them to the given namespace" do
     loader.push_dir(Pathname.new("."), namespace: Namespace)
     check_dirs
+  end
+
+  test "there can be several root directories, with custom namespaces" do
+    files = [["rd1/x.rb", "X = 1"], ["rd2/y.rb", "Y = 1"], ["rd3/z.rb", "Z = 1"]]
+    with_setup(files, namespace: Namespace) do
+      check_dirs
+    end
+  end
+
+  test "there can be several root directories, with custom namespaces, some of them may be ignored" do
+    files = [["rd1/x.rb", "X = 1"], ["rd2/y.rb", "Y = 1"], ["rd3/z.rb", "Z = 1"]]
+    with_files(files) do
+      loader.push_dir("rd1", namespace: Namespace)
+      loader.push_dir("rd2", namespace: Namespace)
+      loader.push_dir("rd3", namespace: Namespace)
+      loader.ignore("rd2")
+      check_dirs
+    end
   end
 
   test "raises on non-existing directories" do